refactor TeamTable to utilize EntityTable for improved structure
